+++ core/src/napi/callback_info.cc
@@ -39,7 +39,7 @@ void CallbackInfo::AddValue(const std::shared_ptr<CtxValue>& value) {
 }
 
 std::shared_ptr<CtxValue> CallbackInfo::operator[](int index) const {
-  if (index < 0 || index >= Length()) {
+  if (index < 0 || static_cast<size_t>(index) >= Length()) {
     return nullptr;
   }
   return values_[index];
